(mod/components)moving components in sub-directories
