+++ src/components/Bars/Sidebar/Dialog.js
