+++ src/components/Logos/MainLogo.js
